requirements/base.txt
celery
flower
redis
bunnet
pydantic < 2.0

# for .env value
python-dotenv

# utils
orjson
tqdm

# telegram
pyrogram
tgcrypto


langdetect
jsonpath_ng
bson
pymongo
celery-redbeat
